+++ include/El/blas_like/level2/QuasiTrsv.hpp
@@ -1,0 +1,63 @@
+#pragma once
+/*
+   Copyright (c) 2009-2016, Jack Poulson
+   All rights reserved.
+
+   This file is part of Elemental and is under the BSD 2-Clause License, 
+   which can be found in the LICENSE file in the root directory, or at 
+   http://opensource.org/licenses/BSD-2-Clause
+*/
+#include <El/core.hpp>
+#include <El/blas_like/level2/QuasiTrsv/LN.hpp>
+#include <El/blas_like/level2/QuasiTrsv/LT.hpp>
+#include <El/blas_like/level2/QuasiTrsv/UN.hpp>
+#include <El/blas_like/level2/QuasiTrsv/UT.hpp>
+
+namespace El {
+
+template<typename F>
+void QuasiTrsv
+( UpperOrLower uplo, Orientation orientation, const Matrix<F>& A, Matrix<F>& x, 
+  bool checkIfSingular )
+{
+    DEBUG_ONLY(CSE cse("QuasiTrsv"))
+    if( uplo == LOWER )
+    {
+        if( orientation == NORMAL )
+            quasitrsv::LN( A, x, checkIfSingular );
+        else
+            quasitrsv::LT( orientation, A, x, checkIfSingular );
+    }
+    else
+    {
+        if( orientation == NORMAL )
+            quasitrsv::UN( A, x, checkIfSingular );
+        else
+            quasitrsv::UT( orientation, A, x, checkIfSingular );
+    }
+}
+
+template<typename F>
+void QuasiTrsv
+( UpperOrLower uplo, Orientation orientation, 
+  const ElementalMatrix<F>& A, ElementalMatrix<F>& x, 
+  bool checkIfSingular )
+{
+    DEBUG_ONLY(CSE cse("QuasiTrsv"))
+    if( uplo == LOWER )
+    {
+        if( orientation == NORMAL )
+            quasitrsv::LN( A, x, checkIfSingular );
+        else
+            quasitrsv::LT( orientation, A, x, checkIfSingular );
+    }
+    else
+    {
+        if( orientation == NORMAL )
+            quasitrsv::UN( A, x, checkIfSingular );
+        else
+            quasitrsv::UT( orientation, A, x, checkIfSingular );
+    }
+}
+
+} // namespace El
