+++ include/El/blas_like/level2/Trsv/LT.hpp
@@ -1,3 +1,4 @@
+#pragma once
 /*
    Copyright (c) 2009-2016, Jack Poulson
    All rights reserved.
@@ -12,81 +13,82 @@ namespace trsv {
 
 template<typename F>
 inline void
-UT
+LT
 ( Orientation orientation,
   UnitOrNonUnit diag, 
-  const AbstractDistMatrix<F>& UPre,
+  const AbstractDistMatrix<F>& LPre,
         AbstractDistMatrix<F>& xPre )
 {
     DEBUG_ONLY(
-      CSE cse("trsv::UT");
+      CSE cse("trsv::LT");
       if( orientation == NORMAL )
           LogicError("Expected a (conjugate-)transpose option");
-      AssertSameGrids( UPre, xPre );
-      if( UPre.Height() != UPre.Width() )
-          LogicError("U must be square");
+      AssertSameGrids( LPre, xPre );
+      if( LPre.Height() != LPre.Width() )
+          LogicError("L must be square");
       if( xPre.Width() != 1 && xPre.Height() != 1 )
           LogicError("x must be a vector");
       const Int xLength =
           ( xPre.Width() == 1 ? xPre.Height() : xPre.Width() );
-      if( UPre.Width() != xLength )
+      if( LPre.Width() != xLength )
           LogicError("Nonconformal");
     )
-    const Int m = UPre.Height();
+    const Int m = LPre.Height();
     const Int bsize = Blocksize();
-    const Grid& g = UPre.Grid();
+    const Int kLast = LastOffset( m, bsize );
+    const Grid& g = LPre.Grid();
 
-    DistMatrixReadProxy<F,F,MC,MR> UProx( UPre );
+    DistMatrixReadProxy<F,F,MC,MR> LProx( LPre );
     DistMatrixReadWriteProxy<F,F,MC,MR> xProx( xPre );
-    auto& U = UProx.GetLocked();
+    auto& L = LProx.GetLocked();
     auto& x = xProx.Get();
 
     // Matrix views 
-    DistMatrix<F> U11(g), U12(g), x1(g);
+    DistMatrix<F> L10(g), L11(g), x1(g);
 
     // Temporary distributions
-    DistMatrix<F,STAR,STAR> U11_STAR_STAR(g), x1_STAR_STAR(g);
+    DistMatrix<F,STAR,STAR> L11_STAR_STAR(g), x1_STAR_STAR(g);
 
     if( x.Width() == 1 )
     {
-        DistMatrix<F,MC,  STAR> x1_MC_STAR(g);
-        DistMatrix<F,MR,  MC  > z1_MR_MC(g);
-        DistMatrix<F,MR,  STAR> z_MR_STAR(g);
+        DistMatrix<F,MC,STAR> x1_MC_STAR(g);
+        DistMatrix<F,MR,MC  > z1_MR_MC(g);
+        DistMatrix<F,MR,STAR> z_MR_STAR(g);
 
         // Views of z[MR,* ]
-        DistMatrix<F,MR,STAR> z1_MR_STAR(g), z2_MR_STAR(g);
+        DistMatrix<F,MR,STAR> z0_MR_STAR(g), z1_MR_STAR(g);
 
-        z_MR_STAR.AlignWith( U );
+        z_MR_STAR.AlignWith( L );
         Zeros( z_MR_STAR, m, 1 );
 
-        for( Int k=0; k<m; k+=bsize )
+        for( Int k=kLast; k>=0; k-=bsize )
         {
             const Int nb = Min(bsize,m-k);
 
-            LockedView( U11, U, IR(k,k+nb), IR(k,k+nb) );
-            LockedView( U12, U, IR(k,k+nb), IR(k+nb,m) );
+            LockedView( L10, L, IR(k,k+nb), IR(0,k) );
+            LockedView( L11, L, IR(k,k+nb), IR(k,k+nb) );
 
             View( x1, x, IR(k,k+nb), ALL );
 
+            View( z0_MR_STAR, z_MR_STAR, IR(0,k), ALL );
             View( z1_MR_STAR, z_MR_STAR, IR(k,k+nb), ALL );
-            View( z2_MR_STAR, z_MR_STAR, IR(k+nb,m), ALL );
 
-            if( k != 0 )
+            if( k+nb != m )
             {
                 Contract( z1_MR_STAR, z1_MR_MC );
                 x1 += z1_MR_MC;
             }
 
             x1_STAR_STAR = x1;
-            U11_STAR_STAR = U11;
+            L11_STAR_STAR = L11;
             Trsv
-            ( UPPER, orientation, diag,
-              U11_STAR_STAR.LockedMatrix(), x1_STAR_STAR.Matrix() );
+            ( LOWER, orientation, diag,
+              L11_STAR_STAR.LockedMatrix(), x1_STAR_STAR.Matrix() );
             x1 = x1_STAR_STAR;
 
-            x1_MC_STAR.AlignWith( U12 );
+            x1_MC_STAR.AlignWith( L10 );
             x1_MC_STAR = x1_STAR_STAR;
-            LocalGemv( orientation, F(-1), U12, x1_MC_STAR, F(1), z2_MR_STAR );
+            LocalGemv( orientation, F(-1), L10, x1_MC_STAR, F(1), z0_MR_STAR );
         }
     }
     else
@@ -94,37 +96,37 @@ UT
         DistMatrix<F,STAR,MC> x1_STAR_MC(g);
         DistMatrix<F,STAR,MR> z_STAR_MR(g);
 
-        // Views of z[* ,MR]
-        DistMatrix<F,STAR,MR> z1_STAR_MR(g), z2_STAR_MR(g);
+        // Views of z[* ,MR], which will store updates to x
+        DistMatrix<F,STAR,MR> z0_STAR_MR(g), z1_STAR_MR(g);
 
-        z_STAR_MR.AlignWith( U );
+        z_STAR_MR.AlignWith( L );
         Zeros( z_STAR_MR, 1, m );
 
-        for( Int k=0; k<m; k+=bsize )
+        for( Int k=kLast; k>=0; k-=bsize )
         {
             const Int nb = Min(bsize,m-k);
 
-            LockedView( U11, U, IR(k,k+nb), IR(k,k+nb) );
-            LockedView( U12, U, IR(k,k+nb), IR(k+nb,m) );
+            LockedView( L10, L, IR(k,k+nb), IR(0,k) );
+            LockedView( L11, L, IR(k,k+nb), IR(k,k+nb) );
 
             View( x1, x, ALL, IR(k,k+nb) );
 
+            View( z0_STAR_MR, z_STAR_MR, ALL, IR(0,k) );
             View( z1_STAR_MR, z_STAR_MR, ALL, IR(k,k+nb) );
-            View( z2_STAR_MR, z_STAR_MR, ALL, IR(k+nb,m) );
 
-            if( k != 0 )
+            if( k+nb != m )
                 AxpyContract( F(1), z1_STAR_MR, x1 );
 
             x1_STAR_STAR = x1;
-            U11_STAR_STAR = U11;
+            L11_STAR_STAR = L11;
             Trsv
-            ( UPPER, orientation, diag,
-              U11_STAR_STAR.LockedMatrix(), x1_STAR_STAR.Matrix() );
+            ( LOWER, orientation, diag,
+              L11_STAR_STAR.LockedMatrix(), x1_STAR_STAR.Matrix() );
             x1 = x1_STAR_STAR;
 
-            x1_STAR_MC.AlignWith( U12 );
+            x1_STAR_MC.AlignWith( L10 );
             x1_STAR_MC = x1_STAR_STAR;
-            LocalGemv( orientation, F(-1), U12, x1_STAR_MC, F(1), z2_STAR_MR );
+            LocalGemv( orientation, F(-1), L10, x1_STAR_MC, F(1), z0_STAR_MR );
         }
     }
 }
