+++ include/El/blas_like/level2/Trsv/UN.hpp
@@ -1,3 +1,4 @@
+#pragma once
 /*
    Copyright (c) 2009-2016, Jack Poulson
    All rights reserved.
@@ -12,37 +13,38 @@ namespace trsv {
 
 template<typename F>
 inline void
-LN
+UN
 ( UnitOrNonUnit diag, 
-  const AbstractDistMatrix<F>& LPre,
+  const AbstractDistMatrix<F>& UPre,
         AbstractDistMatrix<F>& xPre )
 {
     DEBUG_ONLY(
-      CSE cse("trsv::LN");
-      AssertSameGrids( LPre, xPre );
-      if( LPre.Height() != LPre.Width() )
-          LogicError("L must be square");
+      CSE cse("trsv::UN");
+      AssertSameGrids( UPre, xPre );
+      if( UPre.Height() != UPre.Width() )
+          LogicError("U must be square");
       if( xPre.Width() != 1 && xPre.Height() != 1 )
           LogicError("x must be a vector");
-      const Int xLength = 
+      const Int xLength =
           ( xPre.Width() == 1 ? xPre.Height() : xPre.Width() );
-      if( LPre.Width() != xLength )
+      if( UPre.Width() != xLength )
           LogicError("Nonconformal");
     )
-    const Int m = LPre.Height();
+    const Int m = UPre.Height();
     const Int bsize = Blocksize();
-    const Grid& g = LPre.Grid();
+    const Int kLast = LastOffset( m, bsize );
+    const Grid& g = UPre.Grid();
 
-    DistMatrixReadProxy<F,F,MC,MR> LProx( LPre );
+    DistMatrixReadProxy<F,F,MC,MR> UProx( UPre );
     DistMatrixReadWriteProxy<F,F,MC,MR> xProx( xPre );
-    auto& L = LProx.GetLocked();
+    auto& U = UProx.GetLocked();
     auto& x = xProx.Get();
 
-    // Matrix views 
-    DistMatrix<F> L11(g), L21(g), x1(g);
+    // Matrix views
+    DistMatrix<F> U01(g), U11(g), x1(g);
 
     // Temporary distributions
-    DistMatrix<F,STAR,STAR> L11_STAR_STAR(g), x1_STAR_STAR(g);
+    DistMatrix<F,STAR,STAR> U11_STAR_STAR(g), x1_STAR_STAR(g);
 
     if( x.Width() == 1 )
     {
@@ -50,36 +52,36 @@ LN
         DistMatrix<F,MC,STAR> z_MC_STAR(g);
 
         // Views of z[MC,* ], which will store updates to x
-        DistMatrix<F,MC,STAR> z1_MC_STAR(g), z2_MC_STAR(g);
+        DistMatrix<F,MC,STAR> z0_MC_STAR(g), z1_MC_STAR(g);
 
-        z_MC_STAR.AlignWith( L );
+        z_MC_STAR.AlignWith( U );
         Zeros( z_MC_STAR, m, 1 );
 
-        for( Int k=0; k<m; k+=bsize )
+        for( Int k=kLast; k>=0; k-=bsize )
         {
             const Int nb = Min(bsize,m-k);
 
-            LockedView( L11, L, IR(k,k+nb), IR(k,k+nb) );
-            LockedView( L21, L, IR(k+nb,m), IR(k,k+nb) );
+            LockedView( U01, U, IR(0,k), IR(k,k+nb) );
+            LockedView( U11, U, IR(k,k+nb), IR(k,k+nb) );
 
             View( x1, x, IR(k,k+nb), ALL );
 
+            View( z0_MC_STAR, z_MC_STAR, IR(0,k), ALL );
             View( z1_MC_STAR, z_MC_STAR, IR(k,k+nb), ALL );
-            View( z2_MC_STAR, z_MC_STAR, IR(k+nb,m), ALL );
 
-            if( k != 0 )
+            if( k+nb != m )
                 AxpyContract( F(1), z1_MC_STAR, x1 );
 
             x1_STAR_STAR = x1;
-            L11_STAR_STAR = L11;
+            U11_STAR_STAR = U11;
             Trsv
-            ( LOWER, NORMAL, diag,
-              L11_STAR_STAR.LockedMatrix(), x1_STAR_STAR.Matrix() );
+            ( UPPER, NORMAL, diag,
+              U11_STAR_STAR.LockedMatrix(), x1_STAR_STAR.Matrix() );
             x1 = x1_STAR_STAR;
 
-            x1_MR_STAR.AlignWith( L21 );
+            x1_MR_STAR.AlignWith( U01 );
             x1_MR_STAR = x1_STAR_STAR;
-            LocalGemv( NORMAL, F(-1), L21, x1_MR_STAR, F(1), z2_MC_STAR );
+            LocalGemv( NORMAL, F(-1), U01, x1_MR_STAR, F(1), z0_MC_STAR );
         }
     }
     else
@@ -89,39 +91,39 @@ LN
         DistMatrix<F,STAR,MC> z_STAR_MC(g);
 
         // Views of z[* ,MC]
-        DistMatrix<F,STAR,MC> z1_STAR_MC(g), z2_STAR_MC(g);
+        DistMatrix<F,STAR,MC> z0_STAR_MC(g), z1_STAR_MC(g);
 
-        z_STAR_MC.AlignWith( L );
+        z_STAR_MC.AlignWith( U );
         Zeros( z_STAR_MC, 1, m );
 
-        for( Int k=0; k<m; k+=bsize )
+        for( Int k=kLast; k>=0; k-=bsize )
         {
             const Int nb = Min(bsize,m-k);
 
-            LockedView( L11, L, IR(k,k+nb), IR(k,k+nb) );
-            LockedView( L21, L, IR(k+nb,m), IR(k,k+nb) );
+            LockedView( U01, U, IR(0,k), IR(k,k+nb) );
+            LockedView( U11, U, IR(k,k+nb), IR(k,k+nb) );
 
             View( x1, x, ALL, IR(k,k+nb) );
 
+            View( z0_STAR_MC, z_STAR_MC, ALL, IR(0,k) );
             View( z1_STAR_MC, z_STAR_MC, ALL, IR(k,k+nb) );
-            View( z2_STAR_MC, z_STAR_MC, ALL, IR(k+nb,m) );
 
-            if( k != 0 )
+            if( k+nb != m )
             {
                 Contract( z1_STAR_MC, z1_MR_MC );
                 x1 += z1_MR_MC;
             }
 
             x1_STAR_STAR = x1;
-            L11_STAR_STAR = L11;
+            U11_STAR_STAR = U11;
             Trsv
-            ( LOWER, NORMAL, diag,
-              L11_STAR_STAR.LockedMatrix(), x1_STAR_STAR.Matrix() );
+            ( UPPER, NORMAL, diag,
+              U11_STAR_STAR.LockedMatrix(), x1_STAR_STAR.Matrix() );
             x1 = x1_STAR_STAR;
 
-            x1_STAR_MR.AlignWith( L21 );
+            x1_STAR_MR.AlignWith( U01 );
             x1_STAR_MR = x1_STAR_STAR;
-            LocalGemv( NORMAL, F(-1), L21, x1_STAR_MR, F(1), z2_STAR_MC );
+            LocalGemv( NORMAL, F(-1), U01, x1_STAR_MR, F(1), z0_STAR_MC );
         }
     }
 }
