+++ include/El/core.hpp
@@ -85,30 +85,30 @@ typedef __float128 Quad;
 } // namespace El
 
 // Declare the intertwined core parts of our library
-#include "El/core/imports/valgrind.hpp"
-#include "El/core/imports/omp.hpp"
-#include "El/core/imports/qd.hpp"
-#include "El/core/imports/mpc.hpp"
-#include "El/core/Memory.hpp"
-#include "El/core/Element/decl.hpp"
-#include "El/core/types.hpp"
-#include "El/core/Serialize.hpp"
-#include "El/core/imports/mpi.hpp"
-#include "El/core/imports/choice.hpp"
-#include "El/core/imports/mpi_choice.hpp"
-#include "El/core/environment/decl.hpp"
+#include <El/core/imports/valgrind.hpp>
+#include <El/core/imports/omp.hpp>
+#include <El/core/imports/qd.hpp>
+#include <El/core/imports/mpc.hpp>
+#include <El/core/Memory.hpp>
+#include <El/core/Element/decl.hpp>
+#include <El/core/types.hpp>
+#include <El/core/Serialize.hpp>
+#include <El/core/imports/mpi.hpp>
+#include <El/core/imports/choice.hpp>
+#include <El/core/imports/mpi_choice.hpp>
+#include <El/core/environment/decl.hpp>
 
-#include "El/core/Timer.hpp"
-#include "El/core/indexing/decl.hpp"
-#include "El/core/imports/blas.hpp"
-#include "El/core/imports/lapack.hpp"
-#include "El/core/imports/flame.hpp"
-#include "El/core/imports/mkl.hpp"
-#include "El/core/imports/openblas.hpp"
-#include "El/core/imports/pmrrr.hpp"
-#include "El/core/imports/scalapack.hpp"
+#include <El/core/Timer.hpp>
+#include <El/core/indexing/decl.hpp>
+#include <El/core/imports/blas.hpp>
+#include <El/core/imports/lapack.hpp>
+#include <El/core/imports/flame.hpp>
+#include <El/core/imports/mkl.hpp>
+#include <El/core/imports/openblas.hpp>
+#include <El/core/imports/pmrrr.hpp>
+#include <El/core/imports/scalapack.hpp>
 
-#include "El/core/limits.hpp"
+#include <El/core/limits.hpp>
 
 namespace El {
 
@@ -124,31 +124,30 @@ class DistMatrix;
 
 } // namespace El
 
-#include "El/core/Matrix.hpp"
-#include "El/core/Grid.hpp"
-#include "El/core/DistMatrix.hpp"
-#include "El/core/Proxy.hpp"
+#include <El/core/Matrix.hpp>
+#include <El/core/Grid.hpp>
+#include <El/core/DistMatrix.hpp>
+#include <El/core/Proxy.hpp>
 
 // Implement the intertwined parts of the library
-#include "El/core/Element/impl.hpp"
-#include "El/core/environment/impl.hpp"
-#include "El/core/indexing/impl.hpp"
+#include <El/core/Element/impl.hpp>
+#include <El/core/environment/impl.hpp>
+#include <El/core/indexing/impl.hpp>
 
 // Declare and implement the decoupled parts of the core of the library
 // (perhaps these should be moved into their own directory?)
-#include "El/core/View.hpp"
-#include "El/core/FlamePart.hpp"
-#include "El/core/random/decl.hpp"
-#include "El/core/random/impl.hpp"
+#include <El/core/View.hpp>
+#include <El/core/FlamePart.hpp>
+#include <El/core/random/decl.hpp>
+#include <El/core/random/impl.hpp>
 
-#include "El/core/Graph.hpp"
+#include <El/core/Graph.hpp>
 // TODO: Sequential map
-//#include "El/core/Map.hpp"
-#include "El/core/SparseMatrix.hpp"
-
-#include "El/core/DistGraph.hpp"
-#include "El/core/DistMap.hpp"
-#include "El/core/DistMultiVec.hpp"
-#include "El/core/DistSparseMatrix.hpp"
+//#include <El/core/Map.hpp>
+#include <El/core/SparseMatrix.hpp>
+#include <El/core/DistGraph.hpp>
+#include <El/core/DistMap.hpp>
+#include <El/core/DistMultiVec.hpp>
+#include <El/core/DistSparseMatrix.hpp>
 
 #endif // ifndef EL_CORE_HPP
