+++ src/blas_like/level1/Full.cpp
@@ -6,7 +6,10 @@
    which can be found in the LICENSE file in the root directory, or at 
    http://opensource.org/licenses/BSD-2-Clause
 */
-#include "El.hpp"
+#include <El/core/Matrix.hpp>
+#include <El/core/SparseMatrix.hpp>
+#include <El/blas_like/level1/Copy.hpp>
+#include <El/blas_like/level1.hpp>
 
 namespace El {
