+++ src/blas_like/level2/Geru.cpp
@@ -6,123 +6,8 @@
    which can be found in the LICENSE file in the root directory, or at 
    http://opensource.org/licenses/BSD-2-Clause
 */
-#include "El.hpp"
-
+#include <El/blas_like/level2/Geru.hpp>
 namespace El {
-
-template<typename T>
-void Geru
-( T alpha,
-  const Matrix<T>& x,
-  const Matrix<T>& y,
-        Matrix<T>& A )
-{
-    DEBUG_ONLY(
-      CSE cse("Geru");
-      if( ( x.Height() != 1 && x.Width() != 1 ) ||
-          ( y.Height() != 1 && y.Width() != 1 ) )
-          LogicError("x and y must be vectors");
-      const Int xLength = ( x.Width()==1 ? x.Height() : x.Width() );
-      const Int yLength = ( y.Width()==1 ? y.Height() : y.Width() );
-      if( xLength != A.Height() || yLength != A.Width() )
-          LogicError("Nonconformal Geru");
-    )
-    const Int m = A.Height();
-    const Int n = A.Width();
-    const Int incx = ( x.Width()==1 ? 1 : x.LDim() );
-    const Int incy = ( y.Width()==1 ? 1 : y.LDim() );
-    blas::Geru
-    ( m, n, alpha, x.LockedBuffer(), incx, y.LockedBuffer(), incy,
-                   A.Buffer(), A.LDim() );
-}
-
-template<typename T>
-void Geru
-( T alpha,
-  const ElementalMatrix<T>& x, 
-  const ElementalMatrix<T>& y,
-        ElementalMatrix<T>& APre )
-{
-    DEBUG_ONLY(
-      CSE cse("Geru");
-      AssertSameGrids( APre, x, y );
-      if( ( x.Width() != 1 && x.Height() != 1 ) ||
-          ( y.Width() != 1 && y.Height() != 1 )   )
-          LogicError("x and y are assumed to be vectors");
-      const Int xLength = ( x.Width()==1 ? x.Height() : x.Width() );
-      const Int yLength = ( y.Width()==1 ? y.Height() : y.Width() );
-      if( APre.Height() != xLength || APre.Width() != yLength )
-          LogicError
-          ("Nonconformal Geru:\n",
-           DimsString(APre,"A"),"\n",DimsString(x,"x"),"\n",
-           DimsString(y,"y"));
-    )
-    const Grid& g = APre.Grid();
-
-    DistMatrixReadWriteProxy<T,T,MC,MR> AProx( APre );
-    auto& A = AProx.Get();
-
-    if( x.Width() == 1 && y.Width() == 1 )
-    {
-        DistMatrix<T,MC,STAR> x_MC_STAR(g);
-        x_MC_STAR.AlignWith( A );
-        x_MC_STAR = x;
-
-        DistMatrix<T,MR,STAR> y_MR_STAR(g);
-        y_MR_STAR.AlignWith( A );
-        y_MR_STAR = y;
-
-        Geru
-        ( alpha, x_MC_STAR.LockedMatrix(),
-                 y_MR_STAR.LockedMatrix(),
-                 A.Matrix() );
-    }
-    else if( x.Width() == 1 )
-    {
-        DistMatrix<T,MC,STAR> x_MC_STAR(g);
-        x_MC_STAR.AlignWith( A );
-        x_MC_STAR = x;
-
-        DistMatrix<T,STAR,MR> y_STAR_MR(g);
-        y_STAR_MR.AlignWith( A );
-        y_STAR_MR = y;
-
-        Geru
-        ( alpha, x_MC_STAR.LockedMatrix(),
-                 y_STAR_MR.LockedMatrix(),
-                 A.Matrix() );
-    }
-    else if( y.Width() == 1 )
-    {
-        DistMatrix<T,STAR,MC> x_STAR_MC(g);
-        x_STAR_MC.AlignWith( A );
-        x_STAR_MC = x;
-
-        DistMatrix<T,MR,STAR> y_MR_STAR(g);
-        y_MR_STAR.AlignWith( A );
-        y_MR_STAR = y;
-
-        Geru
-        ( alpha, x_STAR_MC.LockedMatrix(),
-                 y_MR_STAR.LockedMatrix(),
-                 A.Matrix() );
-    }
-    else
-    {
-        DistMatrix<T,STAR,MC> x_STAR_MC(g);
-        x_STAR_MC.AlignWith( A );
-        x_STAR_MC = x;
-
-        DistMatrix<T,STAR,MR> y_STAR_MR(g);
-        y_STAR_MR.AlignWith( A );
-        y_STAR_MR = y;
-
-        Geru
-        ( alpha, x_STAR_MC.LockedMatrix(),
-                 y_STAR_MR.LockedMatrix(),
-                 A.Matrix() );
-    }
-}
 
 #define PROTO(T) \
   template void Geru \
