+++ src/blas_like/level2/Hemv.cpp
@@ -6,30 +6,9 @@
    which can be found in the LICENSE file in the root directory, or at 
    http://opensource.org/licenses/BSD-2-Clause
 */
-#include "El.hpp"
+#include <El/blas_like/level2/Hemv.hpp>
 
 namespace El {
-
-template<typename T>
-void Hemv
-( UpperOrLower uplo,
-  T alpha, const Matrix<T>& A, const Matrix<T>& x, T beta, Matrix<T>& y )
-{
-    DEBUG_ONLY(CSE cse("Hemv"))
-    Symv( uplo, alpha, A, x, beta, y, true );
-}
-
-template<typename T>
-void Hemv
-( UpperOrLower uplo,
-  T alpha, const ElementalMatrix<T>& A,
-           const ElementalMatrix<T>& x,
-  T beta,        ElementalMatrix<T>& y,
-  const SymvCtrl<T>& ctrl )
-{
-    DEBUG_ONLY(CSE cse("Hemv"))
-    Symv( uplo, alpha, A, x, beta, y, true, ctrl );
-}
 
 #define PROTO(T) \
   template void Hemv \
