+++ src/blas_like/level2/Her2.cpp
@@ -6,28 +6,10 @@
    which can be found in the LICENSE file in the root directory, or at 
    http://opensource.org/licenses/BSD-2-Clause
 */
-#include "El.hpp"
+
+#include <El/blas_like/level2/Her2.hpp>
 
 namespace El {
-
-template<typename T>
-void Her2
-( UpperOrLower uplo,
-  T alpha, const Matrix<T>& x, const Matrix<T>& y, Matrix<T>& A )
-{
-    DEBUG_ONLY(CSE cse("Her2"))
-    Syr2( uplo, T(alpha), x, y, A, true );
-}
-
-template<typename T>
-void Her2
-( UpperOrLower uplo,
-  T alpha, const ElementalMatrix<T>& x, const ElementalMatrix<T>& y,
-                 ElementalMatrix<T>& A )
-{
-    DEBUG_ONLY(CSE cse("Her2"))
-    Syr2( uplo, T(alpha), x, y, A, true );
-}
 
 #define PROTO(T) \
   template void Her2 \
