+++ src/blas_like/level2/Trmv.cpp
@@ -10,33 +10,6 @@
 
 namespace El {
 
-template<typename T>
-void Trmv
-( UpperOrLower uplo, Orientation orientation, UnitOrNonUnit diag,
-  const Matrix<T>& A, Matrix<T>& x )
-{
-    DEBUG_ONLY(
-        CSE cse("Trmv");
-        if( x.Height() != 1 && x.Width() != 1 )
-            LogicError("x must be a vector");
-        if( A.Height() != A.Width() )
-            LogicError("A must be square");
-        const Int xLength = ( x.Width()==1 ? x.Height() : x.Width() );
-        if( xLength != A.Height() )
-            LogicError("x must conform with A");
-    )
-    const char uploChar = UpperOrLowerToChar( uplo );
-    const char transChar = OrientationToChar( orientation );
-    const char diagChar = UnitOrNonUnitToChar( diag );
-    const Int m = A.Height();
-    const Int incx = ( x.Width()==1 ? 1 : x.LDim() );
-    blas::Trmv
-    ( uploChar, transChar, diagChar, m,
-      A.LockedBuffer(), A.LDim(), x.Buffer(), incx );
-}
-
-// TODO: Implement distributed version
-
 #define PROTO(T) \
   template void Trmv \
   ( UpperOrLower uplo, Orientation orientation, UnitOrNonUnit diag, \
